fonll/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "fonll"
version = "0.1.0"
description = ""
authors = ["Alessandro Candido <candido.ale@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.1"
matplotlib = "^3.7.2"
pandas = "^2.1.0"
seaborn = "^0.12.2"
tabulate = "^0.9.0"
click = "^8.1.7"
numpy = "^1.24.0"

[tool.poetry.scripts]
numerical = "fonll.numerical:run"
analytical = "fonll.analytical:run"
compare = "fonll.compare:compare"
